# payment/docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Payment",
        "version": "1.0.0",
...
    },
    "paths": {
        "/gateway": {
            "get": {
                "tags": [
                    "Gateway"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "description": "Gateway Id",
                        "schema": {
                            "minimum": 1,
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "name",
                        "description": "Gateway name",
                        "schema": {
                            "maxLength": 255,
                            "type": "string"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Gateway"
                                    }
                                }
                            }
                        },
                        "description": "Success\n"
                    }
                },
                "summary": "Fetch Getway(s)"
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GatewayUpdateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Gateway"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Gateway"
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Update Gateway"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GatewayCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Gateway"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Gateway"
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Create Gateway"
            }
        },
        "/gateway/payment-transaction-config": {
            "get": {
                "tags": [
                    "Gateway-Payment-Config"
                ],
                "parameters": [
                    {
                        "name": "gatewayId",
                        "description": "Gateway-Id",
                        "schema": {
                            "format": "int64",
                            "type": "integer"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GatewayPaymentConfig"
                                    }
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Fetch Gateway-Payment-Transaction-Config(s)",
                "description": "If you want to filter with any other config params then add them after gatewayId."
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GatewayPaymentConfigUpdateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Gateway-Payment-Config"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GatewayPaymentConfig"
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Update Gateway-Payment-Transaction-Config"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GatewayPaymentConfigCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Gateway-Payment-Config"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GatewayPaymentConfig"
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Create Gateway-Payment-Transaction-Config"
            }
        },
        "/merchant-gateway": {
            "get": {
                "tags": [
                    "Merchant-Gateway"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "description": "Merchant-Gateway Id",
                        "schema": {
                            "minimum": 1,
                            "type": "integer"
                        },
                        "in": "query"
                    },
                    {
                        "name": "merchantId",
                        "description": "Merchant-Id",
                        "schema": {
                            "maxLength": 255,
                            "type": "string"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MerchantGateway"
                                    }
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Fetch Merchant-Gateway(s)"
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MerchantGatewayUpdateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Merchant-Gateway"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MerchantGateway"
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Update Merchant-Gateway"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MerchantGatewayCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Merchant-Gateway"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MerchantGateway"
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Create Merchant-Gateway"
            }
        },
        "/payment-transaction": {
            "get": {
                "tags": [
                    "Payment-Transaction"
                ],
                "parameters": [
                    {
                        "name": "merchantId",
                        "schema": {
                            "maxLength": 255,
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "transactionId",
                        "description": "Payment service generated unique reference id",
                        "schema": {
                            "maxLength": 36,
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "merchantTransactionId",
                        "description": "Merchant provided transaction id",
                        "schema": {
                            "maxLength": 255,
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "gatewayTransactionId",
                        "description": "Payment gateway generated transaction id",
                        "schema": {
                            "maxLength": 255,
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "startDate",
                        "description": "Lower limit on __createdAt__",
                        "schema": {
                            "format": "date",
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "endDate",
                        "description": "Upper limit on __createdAt__",
                        "schema": {
                            "format": "date",
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "offset",
                        "description": "Current record index",
                        "schema": {
                            "minimum": 0,
                            "type": "integer"
                        },
                        "in": "query"
                    },
                    {
                        "name": "limit",
                        "description": "Total record size",
                        "schema": {
                            "minimum": 1,
                            "type": "integer"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentTransaction"
                                    }
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Fetch Payment-Transaction"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentTransactionRequestSchema"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Payment-Transaction"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentTransaction"
                                }
                            }
                        },
                        "description": "For __bKash__ and __codero__ the response are as following.\n\n{\n    \"status\": 200,\n    \"code\": 0,\n    \"title\": \"Success\",\n    \"details\": \"Success\",\n    \"timestamp\": 16157993678,\n    \"data\": [\n        {\n            \"redirectUrl\": \"http://localhost:8081/api/codero/909b5b718a3c2fd1eee7472295f7215671532712d27871ea162c8f8f625c0242cc7c422d\",\n            \"paymentId\": \"014de79e-7f7c-49d4-81c3-d99eda0c7ea7\",\n            \"processedAmount\": \"30.00\",\n            \"processedCurrency\": \"BDT\",\n            \"paymentStatus\": \"INITIATED\"\n        }\n    ]\n}\n\nHere in the data array, you will find an object. In that object you will find a key named __redirectUrl__. You have to load that url in your browser to connect with the desired gateway"
                    }
                },
                "summary": "Create Payment-Transaction",
                "description": "* To accomodate future transaction-types, some fields have been marked as optional. But for most the cases, following optional fields will be required -\n  * merchantTransactionId\n  * accountId\n  * amount\n  * currency\n\n* Additional properties may be added sometimes but they will be transient. These fields will not be save at Payment service end.\n  * ghoori\n    1. redirectUrl - Customer should be redirected to this url"
            }
        },
...
    },
    "components": {
        "schemas": {
            "GatewayUpdateSchema": {
                "description": "Root Type for GatewayUpdateSchema",
                "required": [
                    "id",
                    "updatedBy"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "format": "int64",
                        "description": "Gateway Id",
                        "minimum": 1,
                        "type": "integer"
                    },
                    "prettyName": {
                        "description": "Pretty name for viewing purpose as __name__ field has some restricted characters",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "description": {
                        "description": "A human readable text",
                        "maxLength": 1024,
                        "type": "string"
                    },
                    "status": {
                        "description": "* 0 - Inactive\n* 1 - Active",
                        "type": "integer"
                    },
                    "baseUrl": {
                        "description": "Gateway Base-Url. Ex - http://localhost:3000",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "additionalProperties": {
                        "$ref": "#/components/schemas/AdditionalProperties",
                        "description": "",
                        "maxLength": 255
                    },
                    "updatedBy": {
                        "description": "Any kind of identification",
                        "maxLength": 255,
                        "type": "string"
                    }
                },
                "example": {
                    "id": 1,
                    "prettyName": "Dummy Gateway",
                    "baseUrl": "https://localhost",
                    "status": 1,
                    "additionalProperties": {
                        "key": "value"
                    },
                    "updatedBy": "test"
                }
            },
            "GatewayPaymentConfigUpdateSchema": {
                "description": "Root Type for GatewayPaymentConfigCreateSchema",
                "required": [
                    "configs",
                    "gatewayId",
                    "updatedBy"
                ],
                "type": "object",
                "properties": {
                    "gatewayId": {
                        "format": "int64",
                        "description": "Gateway Id",
                        "minimum": 1,
                        "type": "integer"
                    },
                    "id": {
                        "format": "int64",
                        "description": "Payment Transaction Config Id",
                        "minimum": 1,
                        "type": "integer"
                    },
                    "configs": {
                        "$ref": "#/components/schemas/Configs",
                        "description": "",
                        "maxLength": 255
                    },
                    "updatedBy": {
                        "description": "Any kind of identification who updates this record",
                        "maxLength": 255,
                        "type": "string"
                    }
                },
                "example": {
                    "gatewayId": 1,
                    "id": 1,
                    "configs": {
                        "name": "test"
                    },
                    "updatedBy": "test"
                }
            },
            "GatewayPaymentConfig": {
                "description": "Fields in a Payment-Config is highly depended on Gateway.\n\n| Gateway name | Configs      |\n|--------------|--------------|\n| dummy        | * id         |\n|              | * name       |",
                "example": {
                    "id": "1",
                    "name": "first"
                }
            },
...
            "MerchantGatewayUpdateSchema": {
                "description": "Root Type for MerchantGatewayCreateSchema",
                "required": [
                    "id",
                    "updatedBy"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "format": "int64",
                        "description": "Merchant-Gateway Id",
                        "minimum": 1,
                        "type": "integer"
                    },
                    "merchantId": {
                        "description": "Merchant-Id",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "gatewayId": {
                        "format": "int64",
                        "description": "Gateway-Id",
                        "minimum": 1,
                        "type": "integer"
                    },
                    "status": {
                        "description": "* 0 - Inactive\n* 1 - Active",
                        "type": "integer"
                    },
                    "callbackUrl": {
                        "description": "Merchant Callback-Url. If no callbackUrl is given at __POST /payment-transaction__, then this url will be used.",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "updatedBy": {
                        "description": "Any kind of identification who updates this record",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "additionalProperties": {
                        "description": ""
                    }
                },
                "example": {
                    "id": "1",
                    "merchantId": "2",
                    "gatewayId": 4,
                    "status": 0,
                    "callbackUrl": "http://localhost",
                    "updatedBy": "test",
                    "additionalProperties": {
                        "store_id": "xyz"
                    }
                }
            },
...
            "Gateway": {
                "description": "Root Type for Gateway",
                "type": "object",
                "properties": {
                    "id": {
                        "format": "int64",
                        "description": "System generated unique id",
                        "minimum": 1,
                        "type": "integer"
                    },
                    "name": {
                        "description": "Gateway name.",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "prettyName": {
                        "description": "Pretty name.",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "description": {
                        "description": "A human readable text",
                        "maxLength": 1024,
                        "type": "string"
                    },
                    "status": {
                        "description": "* 0 -Inactive\n* 1 - Active",
                        "type": "integer"
                    },
                    "baseUrl": {
                        "description": "Gateway Base-Url.",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "additionalProperties": {
                        "$ref": "#/components/schemas/AdditionalProperties",
                        "description": ""
                    },
                    "createdBy": {
                        "description": "Any kind of user-identification",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "updatedBy": {
                        "description": "Any kind of user-identification",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "description": "Creation date",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "description": "Last udpate time",
                        "type": "string"
                    }
                },
                "example": {
                    "id": "4",
                    "name": "dummy",
                    "prettyName": "Dummy Gateway",
                    "status": 0,
                    "baseUrl": "https://localhost",
                    "additionalProperties": {
                        "key": "value"
                    },
                    "createdBy": "test",
                    "updatedBy": "test",
                    "createdAt": "2020-10-04T12:12:02.000Z",
                    "updatedAt": "2020-10-04T12:24:50.000Z"
                }
            },
            "GatewayCreateSchema": {
                "description": "Root Type for GatewayCreateSchema",
                "required": [
                    "name",
                    "prettyName",
                    "baseUrl",
                    "createdBy"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "description": "Gateway name. \nNote that,\n* There should be a plugin installed with this name at Payment service end",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "prettyName": {
                        "description": "Pretty name for viewing purpose as __name__ field has some restricted characters",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "description": {
                        "description": "A human readable text",
                        "maxLength": 1024,
                        "type": "string"
                    },
                    "baseUrl": {
                        "description": "Gateway Base-Url. Ex - http://localhost:3000",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "additionalProperties": {
                        "$ref": "#/components/schemas/AdditionalProperties",
                        "description": ""
                    },
                    "createdBy": {
                        "description": "Any kind of identification who creates this record",
                        "maxLength": 255,
                        "type": "string"
                    }
                },
                "example": {
                    "name": "dummy",
                    "prettyName": "Dummy Gateway",
                    "baseUrl": "https://localhost",
                    "additionalProperties": {
                        "checkIntervals": [
                            10,
                            20,
                            40
                        ]
                    },
                    "createdBy": "test"
                }
            },
...
            "GatewayPaymentConfigCreateSchema": {
                "description": "Fields in a Payment-Config Create request is highly depended on Gateway.\n\nFor example, if __dummy__ gateway is being used, then it will expect the following properties - \n* name",
                "required": [
                    "createdBy",
                    "configs",
                    "gatewayId"
                ],
                "type": "object",
                "properties": {
                    "gatewayId": {
                        "format": "int64",
                        "description": "Gateway Id",
                        "minimum": 1,
                        "type": "integer"
                    },
                    "configs": {
                        "$ref": "#/components/schemas/Configs",
                        "description": ""
                    },
                    "createdBy": {
                        "description": "Any kind of identification who creates this record",
                        "maxLength": 255,
                        "type": "string"
                    }
                },
                "example": {
                    "gatewayId": 1,
                    "configs": {
                        "name": "sample"
                    },
                    "createdBy": "test"
                }
            },
...
        }
    },
...
}